Move FsServiceDefLoader to service-loaders dir.
