+++ packages/unmock-core/src/__tests__/fs-service-def-loader.test.ts
@@ -1,5 +1,5 @@
 import * as path from "path";
-import { FsServiceDefLoader } from "../fs-service-def-loader";
+import { FsServiceDefLoader } from "../service-loaders/fs-service-def-loader";
 import { IServiceDef } from "../interfaces";
 
 const RESOURCES_DIR = path.join(__dirname, "__unmock__");
